(fix) Fix reconnect timers for channel and socket (#1)
* (fix) Fix reconnect timers for channel and socket

The timers for the channels were being reset each time, so it would
reconnect every second for each channel. Remove the reset so that the
reconnections back off.

The socket would not reconnect once disconnected, so update the
socket to schedule reconnect timers when it is disconnected.

* Remove old starscream library.

* Re-add Starscream submodule.

Also remove some Cocoapods cruft since we're no longer using Podfiles.

* Move callbacks and bindings to background

* Run socket callbacks on background thread.

The channel callbacks still run on the main thread. For some reason,
running them on the background thread naively does not work.

Also clean out some project cruft.

* Fix bug with callbacks.

Turns out errors were not being handled properly. Switch to executing
callbacks on the next tick instead of running them in the background.

* Remove 0.1 sec delay.

Also fix a redeclaration of deinit.

* Update Starscream to use master branch.

* Remove delay in setting callbacks.

* Version number bump